+++ new-backend/src/modules/collection/exceptions/collection-image/image-load.exception.ts
