+++ new-backend/src/modules/collection/exceptions/collection-image/no-image-provided.exception.ts
