+++ new-backend/src/modules/collection/exceptions/collection-image/too-large-image.exception.ts
