+++ new-backend/src/modules/collection/exceptions/collection-image/wrong-image-format.exception.ts
