+++ new-backend/src/modules/collection/exceptions/favorite-collection/already-favorite-collection.exception.ts
