+++ new-backend/src/modules/collection/exceptions/favorite-collection/not-favorite-collection.exception.ts
