+++ new-backend/src/modules/collection/exceptions/like-collection/collection-already-liked.exception.ts
