+++ new-backend/src/modules/collection/exceptions/like-collection/not-liked.exception.ts
