+++ new-backend/src/modules/collection/exceptions/personal-collection/no-personal-collection.exception.ts
