+++ new-backend/src/modules/collection/exceptions/personal-collection/personal-collection-exists.exception.ts
