+++ new-backend/src/modules/collection/exceptions/wrong-collection-id.exception.ts
@@ -1,6 +1,6 @@
-import { BadRequestException } from '@nestjs/common';
+import { NotFoundException } from '@nestjs/common';
 
-export class WrongCollectionIdException extends BadRequestException {
+export class WrongCollectionIdException extends NotFoundException {
   constructor() {
     super('WRONG_COLLECTION_ID');
   }
